% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example_path}
\alias{example_path}
\title{Dataset-JSON example file path}
\usage{
example_path(name = NULL)
}
\arguments{
\item{name}{The name of Dataset-JSON example. Run \code{\link[=example_path]{example_path()}} to check
available options, e.g. \code{example_path("ae")}.}
}
\value{
A character vector of absolute paths to Dataset-JSON example files.
}
\description{
\code{\link[=example_path]{example_path()}} retrieves the local path to the bundled Dataset-JSON example
files. These files are compressed with xz.
}
\examples{
example_path("ae")

example_path("adsl")

}
