% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncompress_xz.R
\name{unxz}
\alias{unxz}
\title{Uncompress xz}
\usage{
unxz(from, to)
}
\arguments{
\item{from}{A path to a xz file.}

\item{to}{A path to a new file (uncompressed).}
}
\value{
This function is ran for its side effect of creating a new file and
returns its path, i.e. \code{to}.
}
\description{
\code{\link[=unxz]{unxz()}} decompresses a xz file into a new file.
}
\examples{
newfile <- file.path(tempdir(), "ae.json")
unxz(from = example_path("ae"), to = newfile)

}
