% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.json.examples-package.R
\docType{package}
\name{dataset.json.examples-package}
\alias{dataset.json.examples}
\alias{dataset.json.examples-package}
\title{dataset.json.examples: Dataset-JSON Example Files}
\description{
An R data package providing a set of Dataset-JSON files comprising SDTM and ADaM datasets. Examples are originally from: \url{https://github.com/cdisc-org/DataExchange-DatasetJson}.
}
\author{
\strong{Maintainer}: Ramiro Magno \email{rmagno@pattern.institute} (\href{https://orcid.org/0000-0001-5226-3441}{ORCID})

Other contributors:
\itemize{
  \item Pattern Institute [copyright holder, funder]
}

}
\keyword{internal}
